% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicals.R
\name{.set1}
\alias{.set1}
\title{Process Sets}
\usage{
.set1(x, envir)
}
\arguments{
\item{x}{A character string}

\item{envir}{An environment in which to evaluate values}
}
\value{
A data frame with the processed set
}
\description{
This is an internal function and not intended to be used directly.
It processes small sets.
}
\examples{

## ## below is an example that should generate an (informative) error
## extraoperators:::.set1("(-Inf,x)", envir = environment())

z <- max(mtcars$mpg)
extraoperators:::.set1("(-Inf,z)", envir = environment())

extraoperators:::.set1("(-Inf,30)", envir = environment())

## clean up
rm(z)
}
\keyword{internal}
