% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.R
\name{identity_operators}
\alias{identity_operators}
\alias{\%===\%}
\alias{\%!==\%}
\title{Strict Identity Operators}
\usage{
e1 \%===\% e2

e1 \%!==\% e2
}
\arguments{
\item{e1}{An object to compare.}

\item{e2}{An object to compare.}
}
\value{
A single logical value.
}
\description{
These operators are wrappers around \code{identical()} for whole-object
identity checks. They return a single logical value and are not elementwise
comparisons.
}
\examples{

1:3 \%===\% 1:3
1 \%===\% 1L

1:3 \%!==\% 3:1
1 \%!==\% 1L
}
