% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_all.R
\name{logical all}
\alias{logical all}
\alias{\%agele\%}
\alias{\%agel\%}
\alias{\%agle\%}
\alias{\%agl\%}
\alias{\%age\%}
\alias{\%ag\%}
\alias{\%ale\%}
\alias{\%al\%}
\alias{\%ain\%}
\alias{\%a!in\%}
\title{Several ways to evaluate whether all values meet logical conditions including logical range comparison helpers}
\usage{
e1 \%agele\% e2

e1 \%agel\% e2

e1 \%agle\% e2

e1 \%agl\% e2

e1 \%age\% e2

e1 \%ag\% e2

e1 \%ale\% e2

e1 \%al\% e2

e1 \%ain\% e2

e1 \%a!in\% e2
}
\arguments{
\item{e1}{A number of vector to be evaluated}

\item{e2}{A vector of one or two numbers used to denote the
limits for logical comparison.}
}
\value{
A logical value whether all \code{e1} meet the logical conditions.
}
\description{
Several ways to evaluate whether all values meet logical conditions including logical range comparison helpers
}
\examples{

1:5 \%agele\% c(2, 4)
1:5 \%agele\% c(4, 2) # order does not matter uses min / max

1:5 \%agel\% c(2, 4)
1:5 \%agel\% c(4, 2) # order does not matter uses min / max

1:5 \%agle\% c(2, 4)
1:5 \%agle\% c(4, 2) # order does not matter uses min / max

1:5 \%agl\% c(2, 4)
1:5 \%agl\% c(4, 2) # order does not matter uses min / max

1:5 \%age\% 2
1:5 \%age\% 4

1:5 \%ag\% 2
1:5 \%ag\% 4

1:5 \%ale\% 2
1:5 \%ale\% 4

1:5 \%al\% 2
1:5 \%al\% 4

1:5 \%ain\% c(2, 99)
c("jack", "jill", "john", "jane") \%ain\% c("jill", "jane", "bill")

1:5 \%a!in\% c(2, 99)
c("jack", "jill", "john", "jane") \%a!in\% c("jill", "jane", "bill")
}
